organizing stuff, also found a bug on subscribe page
