+++ webapp/webapp/new_urls/subscribe.py
@@ -1,8 +1,9 @@
 from django.urls import path
 from webapp.new_views.subscribe import (
-    success,
+    success, subscribe
 )
 
 urlpatterns = [
+    path("", subscribe, name="subscribe"),
     path("success/", success, name="subscribe_success"),
     ]
