+++ webapp/webapp/urls.py
@@ -26,7 +26,6 @@ urlpatterns = [
     path('create-checkout-session/', views.create_checkout_session, name='create-checkout-session'),
     path('create-customer-portal-session/', views.create_customer_portal_session, name='create-customer-portal-session'),
     path('vocab-test/', views.vocab_test, name='vocab-test'),
-    path('subscribe/', views.subscribe, name='subscribe'),
     path('flag-question/', views.flag_question, name='flag-question'),
     path('vocab-builder/', views.vocab_builder, name='vocab-builder'),
     path('review/', views.review, name='review'),
